Correct incorrect type annotation inside strategy code
This was not causing an error in master because no uses of the
incorrectly annotated attribute (self.executors) were type checked.

This PR introduces a typed dict to represent the (single element) dictionary
used for strategy state.

(A futher tidyup could be to remove that typed dict entirely and replace with
the single value it contains, but this PR is trying to be descriptive of
existing code)

The incorrect type annotation was added in #2712