Tidyup messy join app refactoring PR #2621 (#2861)
PR #2621 didn't remove the user facing join parameter it was intended to,
instead just ignoring it.

That PR #2621 also added a new executors parameter to the join_app decorator,
which is then ignored (as it should be - because join_apps should be forced
to run in the _parsl_internal executor). This PR removes that spurious
parameter.

This will be a breaking change for anyone explicitly specifying the
_parsl_internal decorator to join_app, or for anyone explicitly specifying
join=False to python_app. (Other uses of those parameters are already broken
because those parameters are ignored, and they will now be broken in
different ways)