Add user-pluggable block error handling and a new sliding-window error handler (#2858)
This PR aims to address the limited capabilities of the current simple_error_handler at stopping the Parsl runtime when there are repeated failures. The current system only fails if all jobs fail, which is only indicative of configuration errors or problem with the batch scheduler.

This PR adds new behavior that updates the existing block_error_handler bool variable to take a custom error handler. 

This PR also adds a new windows_error_handler to better handle long running workflows. This handler shuts down if and only if the last N jobs all failed.

Co-authored-by: Ben Clifford <benc@hawaga.org.uk>