Remove execute_wait stdout, stderr optionality
No in-codebase channels return None for stdout or stderr,
and some cluster providers (at least pbspro and slurm) do not
handle the case when stdout is None, instead assuming it will
be a str.

This PR makes those assumptions concrete.

This PR will introduce API breakage for any externally implemented
channels which return None here, although in practice because of
the above, it's likely that use of those channels woudl already
be broken.

However, I am not aware of any externally implemented channels.