Re-enable str-bytes-safe mypy error, and fix uses
These changes are all in the interchange. In most cases, this error is
found when a manager id (a byte sequence) is used in a string substitution,
and to make happy, the format string is modified to indicate that we
really do want to represent that byte sequence as a repr.

The only other place is one rendering of a binary message into a log
string, and the fix is the same.

from benc-mypy branch