Raise configuration error if no executors are specified
I think this is probably the right thing to do.

Breaking change:

There's a corner case where you would only use the implicit
_parsl_internal executor:
eg. if you were only doing join_apps and file staging, and
if you were co-ordinating non-Parsl futures and not running any
python_ or bash_ apps. But I think that is sufficiently obscure, this
check is the right thing to do and if anyone complains, it can be
removed again.