Accept an iterable for executor config
This is a supertype of the previous-commit Sequence, itself a
supertype of List that is current in master.

It subtly emphasises that the collection of iterators will be read
over once by Parsl and not touched again.

This drives a change in the validation call to validate after copying
into an internal structure.