Switch configuration executor List type to Iterable (#2943)
The list of executors in configuration is intended to be read only. Using Iterable for this makes that more explicit, while also allowing other structures that make sense such as tuple or set.

This is part of a move towards using more constrainted sequence types (for example Sequence) rather than List in many places (although the runtime type remains a list) because these can have better behaviour with subclasses of its parameter type: sequences and iterables are covariant, rather than invariant.

This PR rearranges the validation logic a bit so that there is no longer an arbitrary property setter for the executor property. It was always a bug to replace the executor list in a config once Parsl has been configured, and removing this property setter enforces that more strongly.

Co-authored-by: Kevin Hunter Kesling <kevin@globus.org>