Install into virtualenv inside CI
This is motivated by upcoming PR #2923 which adds a RADICAL-Pilot
executor: that executor is able to re-use a submit side virtualenv,
but not able to re-use `--user` level installs.

This should not change any other behaviour of the CI: packages will
be installed at different paths, but anything that was reliant on
those paths being a specific values is suspicious.