Pin release process to Python 3.11 (#2953)
The release process previously used whichever environment the actions/setup-python action chose to use in the 3.x range. Python 3.12 and the current release process are incompatible, and auto-releases broke when actions/setup-python chose to supply Python 3.12. This PR pins the release to Python 3.11, which will remain in a supported state until October 2027. Beyond that, this release process will probably need some re-engineering to work.

Attempts to fix #2934