Update monitoring task_inputs after input Futures are resolved
The monitoring task table stores a column that should store the inputs
argument of a task, after Future resolution.

Prior to this PR:

Although the sending side of the code, DataFlowKernel._send_task_log_info,
is correctly invoked to send that information, it first sends an
earlier representation of that task where inputs are still unresolved
futures - for example:

[<AppFuture at 0x7fca83d22610 state=pending>, <AppFuture at 0x7fca83d7b550 state=pending>, <AppFuture at 0x7fca83d80110 state=pending>]

When a later update is sent, not all columns of the task table are updated;
only a specific list of columns that are expected to change. task_inputs was
not on that list, so even after future resolution, the monitoring task_inputs
column was not updated.

This PR adds task_inputs to the list of columns that are updated on every
task update message.

This PR adds a test which if run without this change results in this test
failure:

```
            result = connection.execute(text("SELECT task_inputs FROM task"))
            (task_inputs, ) = result.first()
>           assert task_inputs == "[" + repr(TOKEN) + "]"
E           AssertionError: assert '[<Future at ...ate=pending>]' == '[54739]'
E             - [54739]
E             + [<Future at 0x7f946e3f9010 state=pending>]

parsl/tests/test_monitoring/test_incomplete_futures.py:66: AssertionError

```