Create thread to handle HTEX monitoring
Created an HTEX manager thread to transfer monitoring messages from a
managed queue, which we can utilize across processes started in fork and
spawn contexts, to the results queue.

We separate the queues so that the results queue does not rely on a
manager process, which causes slower queue operations.