Use spawn for process worker pool (#2983)
Use spawn for process worker pool (#2983)

To avoid issues with forked processes and threads (see #2343), we will
use spawn calls to start HTEX workers.

- Created spawn and fork multiprocessing context objects that we can use
throughout the codebase to increase clarity on how we start processes.

- Migrated away from using the `MacSafeQueue` in the process worker pool
since it was only needed to get the sizes of two queues for a single
debug log statement.

- Replaced `ready_worker_queue` with `ready_worker_count`. The former
was a multiprocessing `Queue` that we only used to track the number of
workers that are ready to receive tasks. The latter is a multiprocessing
`Value`, which is a more appropriate data structure.

- Created an HTEX manager thread to transfer monitoring messages from a
managed queue, which we can utilize across processes started in fork and
spawn contexts, to the results queue. We separate the queues so that the
results queue does not rely on a manager process, which causes slower
queue operations.