Add more info to DependencyError for non-task Futures
In some situations, a dependency in a dependency error does not have a
task ID (becuase it is not an AppFuture) (case 1), or has a task ID from a
different DFK (case 2)

Previously:

Case 1: the dependency error listed a task ID as `None`.

Case 2: the dependency error listed a task ID as if it was that task in the
same DFK as the dependent task.

In both those cases now, the error message renders those "foreign" Futures
using repr.

This PR also tidies up list formatting to remove []-brackets around the list
of dependencies.

Now:

Dependency failure for task 0 with failed dependencies from <Future at 0x7f8ca335ef90 state=finished raised ValueError>

Previously:

Dependency failure for task 0 with failed dependencies from [None]