Add slurmprovider options (#2986)
Adds optional parameters to SlurmProvider for choosing the queue and constraint slurm options. This makes the SlurmProvider easier to work on NERSC systems which uses these options for job submissions. Queue is used to choose the jobs priority (regular, debug, realtime, etc.) and constraint to choose between cpu and gpu nodes on Perlmutter.