Only monitor processes when resource_monitoring_interval > 0
Two kinds of monitoring message originate from workers:

1. task start and end messages which are relatively lightweight, sent by the
task thread at the beginning and end of a task.

2. periodic resource usage message which happen periodically while a task is
running. These are sent from a forked process which monitors the task
process and its descendants. This can be substantially more resource
intensive and degrade workflow performance.

Previously configuration only allow disabling both of this kind of message.
This PR allows the type 1 messages to continue to be sent, allowing users to
see when a task starts executing, without running a per-task periodic
monitoring process.