Add a regression test for broken SlurmProvider repr(), issue #2994 (#2995)
This test checks that repr runs ok for SlurmProvider. That's a method that one might expect would always work, but RepresentationMixin has been a source of exceptions in the past (although with #2996, it would hopefully be less aggressive).

This test does not need a slurm installation, as it does not test any interactions with slurm itself.