Mark radical MPI test (#3001)
A temporary measure to allow testing locally on dev setups that don't have
`radical.pilot` installed and/or working.  Use with:

    pytest parsl/tests/ -k "not radical" ...

For now, this is not _implemented_ anywhere (e.g., not in `Makefile`) so it's
strictly a manual usage.  At some point, the tests are due for an overhaul to
clean up the testing situation.  I won't enumerate the specifics here, but
suffice to say that this is a on-going conversation.