Set default test executor working dir (#3003)
In pursuit of no test detritus, set the default working directory of executors
and the `dfk` to a test-run-specific temporary directory.  For "conforming"
tests, (i.e., those that are aware of the fixtures in `conftest.py` that
attempt to make fresh configurations), this enables auto-cleanup.  The rest
will follow in piecemeal PRs "as we get there."