Split out `make test` into a matrix
Parallelize the tests by using GH's matrix.  There's no need to execute the
tests serially.

This reduces CI build time to ~11m, with the vineex test being the bottleneck.
Accordingly, reduce timeout limit from 60m to 15m, with the expectation that if
tests take longer than that, the *tests* need to be addressed.  In the
semi-rare case of downloads taking too long, the developer in question can run
another attempt in GH's CI UI.