Add .pytest/ to runinfo artifact collection (#3012)
Recent efforts to clean up the test detritus left in the developer's source
directory were unaware of the fact that the `runinfo/` directory is saved in
CI.  To accommodate both goals, refactor the `tmpd_cwd_session` fixture to
preserve the last `N` test runs.  If not specified, default to keeping the last
3 runs.

Meanwhile, tell the upload-artifact step to preserve both directories for the
time being.