Implement tox for test isolation
The main change is to implement tox, a tool for test isoliation.  This is
difficult to do entirely given that some of the tools under test aren't
strictly Python, so let there be some overlap with Make.  In particular Vine
and WorkQ need `/tmp/cctools/bin` in the PATH, in addition to the PYTHONPATH,
so those will most easily be executed via `make`.

Meanwhile, since the recipe logic is now largely in `tox`, cleanup the Makefile
to basically a single stanza.

While here, consolidate project requirements to setup.py.  They're pulled into
setup.py anyway, and one list is as good as another.  More specifically, we had
both `.[test]` and `test-requirements.txt`, nominally as dual points of
authority.  Simplify to exactly one place.

Finally, address an as-yet unidentified issue that appears to be strictly on CI
boxes surrounding the test-session cleanup.