Update JobStatusPoller to not require DFK (#3020)
Currently, the JobStatusPoller and PollItem both take the DataFlowKernel as an argument, when in reality there's very little dependence for these on the DFK. PollItem uses the DKF to get to the monitoring component, which is again optional. This PR converts the DFK to an optional argument.

This PR is primarily to simplify integrating Parsl scaling logic directly with individual executors without involving the DFK which is the direction globus-compute is taking with it's execution engines.