Make RepresentationMixin validation less aggressive (#2996)
Prior to this PR:
RepresentationMixin fails when repr() is called if an implementing class does not store all of the parameters used in its init method.

This leads to a fairly common user-exposed error not caught in CI in Parsl components which are not tested in CI: primarily the provider classes. This immediate work, for example, is motivated by issue #2994 introduced in PR #2986 which is in release 2023.12.11.

This PR changes the behaviour of repr so that it will by default use the string <unrecorded> instead of raising an error at runtime in this situation.

For testing/validation purposes, the validation code is re-enabled when running pytests, so that other pytests (such as PR #2995) can be implemented to fail rather than silently omit parameters.