Create CurveZMQ context classes
`ServerContext` and `ClientContext` replace the standard `zmq.Context`.
They share many commonly used methods, including `term`, `destroy` and,
most importantly, `socket`. The latter applies the necessary certs and
options to each socket object.

A connection requires a `ServerContext` on one end, which validates
clients, and a `ClientContext` on the other, which validates the server.
Certificates are generated via the `create_certificates` function.