Remove use of tmp_path fixture.
PR #2764 introduced `tmp_path` without understanding of the shared FS
expectation for some tests (i.e., those run on site-specific hosts).  PR #2803
fixed this with `tmpd_cwd` which is a temporary directory created inside of the
current working directory -- which is assumed to be on a shared FS.

Also remove use of `tmpdir` fixture.

Fixes: #3039