Add exception type to task failed log messages
This arises from a user interaction where it was unclear that some tasks
were failing and retrying due to an application error; and others were
failing due to lost managers.

The exception is recorded in the monitoring database, if that is used, and
when retries have run out, is returned to the user in the relevant AppFuture
but the exception class was not previously recorded in the log file for
live investigation while a run is still proceeding.

Before:

1706696631.320572 2024-01-31 10:23:51 MainProcess-4072 HTEX-Queue-Management-Thread-139622759765696 parsl.dataflow.dflow:305 handle_exec_update DEBUG: Task 9 try 0 failed

After:

1706696578.454484 2024-01-31 10:22:58 MainProcess-3792 HTEX-Queue-Management-Thread-140175535478464 parsl.dataflow.dflow:305 handle_exec_update INFO: Task 9 try 0 failed with exception of type AppTimeout