Remove unused codepath from HighThroughputExecutor scale_in
The codepath would be used when scale_in is called with:

  force=False
  max_idletime=None

and would pick blocks from the (weirdly sorted - see issue #2120) list of
blocks which are currently idle.

This PR removes that unused codepath, merging the choice of
forced/non-forced scale-in and idle time specification into a single
parameter that indicates "do not scale in blocks that have not been
idle this long".