Make HTEX default launch_cmd available as a global
This supports the use case where a user wants to launch the regular
launch command but inside some kind of wrapper: they can now write

```
  HighThroughputExecutor(
      launch_cmd = "some_wrapper " + HighThroughputExecutor.DEFAULT_LAUNCH_CMD,
      ...
  )
```

which will use the launch command from the current Parsl version.

The alternative is to cut-and-paste the launch command out of the Parsl
source code, which is more fragile.