Make manager->interchange tasks_output protocol always JSON objects
This is to permit imminent changes to this protocol to allow quenching of
task delivery as part of work to drain workers near the end of their
walltime.

Prior to this PR, the protocol used two formats: the first message in the
worker->interchange task_output pipe was a JSON object, the
"registration message".

Subsequent messages were not JSON, but an integer encoded as a byte sequence.
This integer was historically used for two purposes: as a heartbeat
(with value HEARTBEAT_CODE = (2 ** 32) - 1) and as a count of tasks to
request from the interchange.

This latter use was removed in PRs #2588 and #3062, leaving only the
heartbeat use case.

To allow this channel to be used for richer messages, this PR switches the
messages on this stream to always be JSON object, with a 'type'
entry: currently 'registration' or 'heartbeat'.

This is in line with the 'result_package' format used on the
manager->interchange results_outgoing channel, and so aligns with issue
#3022 which talks about unifying those two channels.

Code paths:

This PR removes the reg_flag variable used in registration code, which was
redundant and only used to route control flow from a try block into an
immediately following code block: that second block is merged upwards.

That try block protected/protects json decoding, and is moved higher up
because there is now no need to have a separate path for "expecting JSON"
and "expecting encoded int".

Type checking: This PR adds some additional checking for JSON structure to
report a more useful error when the message is not a dictionary; this drives
mypy to specialise the type of `msg` to dict[Any, Any] which can then not be
used to update a ManagerRecord. This PR adds a # type: ignore with
commentary to address that.

Testing: while registration is implicitly tested by htex tests successfully
running tasks, there isn't any automated testing of manager->interchange
heartbeats, so human review should pay special attention there.