Add end-of-walltime worker draining to htex
After a configured execution period (usually this would be configured to be a
little less than the walltime of the containing batch job), the manager can
request that it be sent no more tasks; it will continue executing task that
it already has, and then when none are left, it will exit.

This is implemented by adding a new 'drain' message to the task_outgoing
manager->interchange channel, which causes a new draining flag to be set
in the interchange; and then a new 'you are drained' message is sent
the other way when the interchange knows that the manager is drained, to
cause the manager to exit.