Fix address field in TaskVine manager config.
Previously it looks like this field was ignored, because it was always
(in the absence of a project name) overriddel by
parsl.addresses.get_any_address. This PR makes that override only happen
when the address field is set to None.

Previously this field had a class level default value of socket.gethostname()
That value was never used, because of the above override, and resulted in the
hostname of the documentation build host being listed as the default value
for this parameter, incorrectly.

This PR removes that default value, allows the address field to be None,
and makes the address field be None by default.

Cross-reference this Debian documentation reproducibility issue where the
presence of socket.gethostname() breaks binary reproducibility:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1063542

Testing:

Prior to this PR, setting address='10.10.10.10' does not break anything, even
though that is not an IP address of my test environment. After this PR,
setting that IP address causes taskvine to hang as I would expect.