test partial scaling-in (the scaling in the comes from one block being idle, not from the whole executor being idle)
this is quite awkward ot test. what can we do?

the current test has min_blocks 2, which means if we launch a single task (pretty much all we can), we'll not be able to distinguish min_blocks scaling from single task scaling. so a separate test perhaps with min_blocks 0?