Update style of scaling test File usage
In this particular test configuration, this doesn't really matter:
all that is being passed around is a path wrapped in an object; no
staging happens; and if staging was configured, this test would usually
not work... so it's arguable even that File shouldn't be used here at
all and just pass around paths.

It is also not necessary to cast an os.PathLike to a str before constructing
a File around it.

Neither of these changes should change the behaviour of this test, but its
probably better that they are in the right style in case someone looks at
them.