Allow a few seconds for the end message usage process to work
Before this PR, the end message process was terminated immediately, which
mean that end messages (in all of my tests) were not sent.

This PR waits 10 seconds for termination before forcing a (now stronger)
kill, which in practice is enough for the end message to be sent.

Reasons for making that timeout any larger are primarily DNS-resolution-time
related - the tracking DNS record has a TTL of only a few hundred seconds,
so it's likely to need re-resolving at end time in any non-trivial run.

In practice on a small pytest run on my laptop, with DNS cached, this new
close procedure takes 20ms.

Fixes issue 2532