Use run_dir from Config for monitoring path
Prior to this PR, the default was hard-coded to "runinfo/" (in PR #2136)
which works most of the time, as most of the time, users do not change
the Config run_dir.

However, PR #3003 sets Config.run_dir for most tests, resulting in a
(mostly silent) monitoring database manager failure. This PR fixes that
by allowing monitoring to see that new path, as well as fixing it for
all other run_dir changing users.

Note that in Config/DFK there are two uses of the "run_dir" : in
Config, that name refers to the root in which .../000/... /NNN/
directories will be created. In the DFK, that name refers to the
numbered config.DFK/NNN/ directory.

Since #3003 (which I bisected back to), monitoring has been mostly
silently failing with a database opening error appearing in
database_manager.log and a missing monitoring.db, but no exception
raised in the workflow process. This PR does not address that
loss-of-failure-info problem.