Remove "anonymous" in relation to usage tracking (#3096)
While this and similar words have been used repeatedly in sellling usage
tracking to end users, the user's IP address (regarded as at least possibly
identifying information under e.g. GDPR, CCPA and HIPAA) is clearly logged on
the receiving side, and for at least that reason it is likely incorrect
to be calling this information collection anonymous.

https://github.com/Parsl/parsl_tracking/blob/2c9e11e991e38100126d28b18a57f52a0cc01aa0/tracking_listener.py#L38

Until PR #2666, an easily reversible encoding of username was transmitted,
suggesting a general culture of insincerity or naivety around anonymisation,
and the onus should now be on the Parsl team to demonstrate a more serious
approach before using this kind of phrasing again.