Remove unused codepath from HighThroughputExecutor scale_in (#3115)
The codepath would be used when scale_in is called with:

  force=False
  max_idletime=None

and would pick blocks from the list of blocks which are currently idle.

This PR removes that unused codepath, merging the choice of
forced/non-forced scale-in and idle time specification into a single
parameter that indicates "do not scale in blocks that have not been
idle this long".

The two remaining cases from force=True/False, max_idletime=None/number
are:


* max_idletime=None, previously: force=True, max_idletime=None

This means that the requested number of blocks should be scaled in, even
if the blocks are not idle. The use case for this path is "parsl is shutting
down, so we need to tidy up everything. If there are tasks still running, we
don't care because terminating in-progress tasks is part of parsl shutdown."

* max_idletime=some time, previousy: force=False, max_idletime=some time

This means that the scaling code has decided to apply downwards pressure on
the number of blocks: there are more blocks than needed. However, this
pressure should not disrupt already running tasks, and it is less urgent to
cancel blocks, because the same call will happen every 5 seconds to keep
applying that pressure.