Rename monitoring port variables to reflect use
The naming of the interchange port suggests that is for the interchange
to send monitoring messages, but it is a general ZMQ listening port used
by at least the DFK too.

There is another general listening port, listening on UDP, used by the
UDPRadio on remote workers.

This PR renames most variables referencing the above connections to have
zmq_ and udp_ prefixes, instead of a mix of hub_ and ic_ prefixes, to
clarify which variables and parameters apply to which of those two
ports.

This clarifies in most of the codebase, for example, that the former
hub_port and hub_port_range variables do not control the same thing in
different ways, but that one is for ZMQ and ones is for UDP.

This PR avoids changing user-facing parameters, so some ambiguosly named
parameters still exist: those should be dealt with by a deprecation
process over a longer time period in a different PR.