Add shebang to exec_parsl_function.py
The script is flagged for installation, so end users may expect to be
able to run it straight from their PATH, without having to specify the
interpreter and the full path to the script.  While the shebang is
missing, the operating system can barely guess what interpreter to use
for running the script in such conditions, and thus would not be able
to run exec_parsl_function.py.

Signed-off-by: Étienne Mollier <emollier@debian.org>