Preserve argument defaults to Sphinx documentation
This change improves Sphinx documentation reproducibility by
preserving argument defaults.

The TaskVineManagerConfig dataclass includes an 'address' attribute
that is set to the value of socket.gethostname() when the class is
loaded.

Meanwhile, the TaskVineExecutor.__init__ method 'manager_config'
argument has a default value of a no-args constructed
TaskVineManagerConfig instance.

When Sphinx builds documentation, by default it will emit a Python
repr() of the manager_config argument, causing the hostname of the
build host to be included.

We can solve that by instructing the Sphinx autodoc extension to
retain the textual representation of argument lists as they are found
in the source code, instead of evaluated and repr'd equivalents.

Signed-off-by: Étienne Mollier <emollier@debian.org>