Make JobStatusPoller parameters mandatory and remove defaults
Making the parameters mandatory and without defaults helps in
debuggability and in code understanding.

Defaults come from closer to the user, and so the defaults removed here
are never used.

To support Parsl users using JobStatusPoller as part of a DFK-free
environment (for example, Globus Compute), this PR leaves the dfk
parameter to have a meaningful default.